Ws device delete returns 500

+++ devicehive-frontend/src/main/java/com/devicehive/resource/impl/DeviceResourceImpl.java
@@ -46,8 +46,8 @@ import java.util.*;
 
 import static com.devicehive.configuration.Constants.*;
 import static com.devicehive.json.strategies.JsonPolicyDef.Policy.DEVICE_PUBLISHED;
-import static javax.servlet.http.HttpServletResponse.SC_NOT_FOUND;
 import static javax.ws.rs.core.Response.Status.BAD_REQUEST;
+import static javax.ws.rs.core.Response.Status.NOT_FOUND;
 import static javax.ws.rs.core.Response.Status.NO_CONTENT;
 
 /**
@@ -65,12 +65,11 @@ public class DeviceResourceImpl implements DeviceResource {
      */
     @Override
     public void list(String name, String namePattern, Long networkId, String networkName,
-                     String sortField, String sortOrderSt, Integer take,
+                     String sortField, String sortOrder, Integer take,
                      Integer skip, @Suspended final AsyncResponse asyncResponse) {
 
         logger.debug("Device list requested");
 
-        boolean sortOrder = SortOrder.parse(sortOrderSt);
         if (sortField != null
                 && !NAME.equalsIgnoreCase(sortField)
                 && !STATUS.equalsIgnoreCase(sortField)
@@ -96,7 +95,7 @@ public class DeviceResourceImpl implements DeviceResource {
             request.setNetworkId(networkId);
             request.setNetworkName(networkName);
             request.setSortField(sortField);
-            request.setSortOrderAsc(sortOrder);
+            request.setSortOrder(sortOrder);
             request.setTake(take);
             request.setSkip(skip);
             request.setPrincipal(principal);
@@ -141,7 +140,9 @@ public class DeviceResourceImpl implements DeviceResource {
 
         if (device == null) {
             logger.error("device/get proceed with error. No Device with Device ID = {} found.", deviceId);
-            throw new HiveException(String.format(Messages.DEVICE_NOT_FOUND, deviceId), SC_NOT_FOUND);
+            ErrorResponse errorResponseEntity = new ErrorResponse(NOT_FOUND.getStatusCode(),
+                    String.format(Messages.DEVICE_NOT_FOUND, deviceId));
+            return ResponseFactory.response(NOT_FOUND, errorResponseEntity);
         }
 
         logger.debug("Device get proceed successfully. Device ID: {}", deviceId);
@@ -153,10 +154,19 @@ public class DeviceResourceImpl implements DeviceResource {
      */
     @Override
     public Response delete(String deviceId) {
+        if (deviceId == null) {
+            logger.error("device/get proceed with error. Device ID should be provided.");
+            ErrorResponse errorResponseEntity = new ErrorResponse(BAD_REQUEST.getStatusCode(),
+                    Messages.DEVICE_ID_REQUIRED);
+            return ResponseFactory.response(BAD_REQUEST, errorResponseEntity);
+        }
+        
         boolean isDeviceDeleted = deviceService.deleteDevice(deviceId);
         if (!isDeviceDeleted) {
             logger.error("Delete device proceed with error. No Device with Device ID = {} found.", deviceId);
-            throw new HiveException(String.format(Messages.DEVICE_NOT_FOUND, deviceId), SC_NOT_FOUND);
+            ErrorResponse errorResponseEntity = new ErrorResponse(NOT_FOUND.getStatusCode(),
+                    String.format(Messages.DEVICE_NOT_FOUND, deviceId));
+            return ResponseFactory.response(NOT_FOUND, errorResponseEntity);
         }
         
         logger.debug("Device with id = {} is deleted", deviceId);
